+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Notes for Studio plugin
 
+## Unreleased
+
+- Changed field handle of matrix field for value4value items to `valueRecipient` because `recipient` field handle is taken and used in some matrix sub fields in fresh Craft5 projects.
+
 ## 2.0.0-alpha.2 - 2024-05-07
 
 - Fixed a bug where episode settings, general settings and import episodes actions were available for draft podcasts.